[Upg] start migration to silex 2
