+++ composer.json
@@ -6,7 +6,7 @@
     "authors": [
         {
             "name": "Mathieu Rochette",
-            "email": "mathieu@texthtml.net",
+            "email": "mathieu@rochette.cc",
             "homepage": "https://mathieu.rochette.cc"
         },
         {
@@ -19,11 +19,11 @@
     {
         "php": ">=5.4.0",
         "doctrine/orm": "~2.4",
-        "silex/silex": "~1.2",
-        "dflydev/doctrine-orm-service-provider": "~1.0",
-        "saxulum/saxulum-doctrine-orm-manager-registry-provider": "~2.0",
-        "cilex/cilex": "~1.1",
-        "cypresslab/gitelephant": "dev-develop"
+        "silex/silex": "~2.0@dev",
+        "dflydev/doctrine-orm-service-provider": "dev-master@dev",
+        "saxulum/saxulum-doctrine-orm-manager-registry-provider": "dev-master@dev",
+        "cypresslab/gitelephant": "dev-develop",
+        "gitory/pimple-cli": "dev-master"
     },
     "require-dev":
     {
